Refactored classes in PCL project and updated README.
