+++ DigiReceipt/DigiReceipt/DigiReceipt/DigiReceipt.csproj
@@ -43,7 +43,7 @@
     <Compile Include="Data\Receipt.cs" />
     <Compile Include="Models\ReceiptsModel.cs" />
     <Compile Include="Models\ReceiptModel.cs" />
-    <Compile Include="BooleanConverter.cs" />
+    <Compile Include="Converters\BooleanConverter.cs" />
     <Compile Include="Resx\AppResources.Designer.cs">
       <AutoGen>True</AutoGen>
       <DesignTime>True</DesignTime>
@@ -63,7 +63,7 @@
       <DependentUpon>Login.xaml</DependentUpon>
     </Compile>
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="TranslateExtension.cs" />
+    <Compile Include="Markup\TranslateExtension.cs" />
     <Compile Include="Views\ViewReceipt.xaml.cs">
       <DependentUpon>ViewReceipt.xaml</DependentUpon>
     </Compile>
@@ -183,6 +183,7 @@
     <None Include="app.config" />
     <None Include="packages.config" />
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildExtensionsPath32)\Microsoft\Portable\$(TargetFrameworkVersion)\Microsoft.Portable.CSharp.targets" />
   <Import Project="..\..\packages\Xamarin.Forms.2.4.0.38779\build\portable-win+net45+wp80+win81+wpa81\Xamarin.Forms.targets" Condition="Exists('..\..\packages\Xamarin.Forms.2.4.0.38779\build\portable-win+net45+wp80+win81+wpa81\Xamarin.Forms.targets')" />
   <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
