+++ DigiReceipt/DigiReceipt/DigiReceipt/Views/ViewReceipts.xaml
@@ -1,24 +1,24 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
 <ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             xmlns:local="clr-namespace:DigiReceipt"
+             xmlns:markup="clr-namespace:DigiReceipt.Markup"
              xmlns:sys="clr-namespace:System;assembly=mscorlib"
              xmlns:behaviours="clr-namespace:DigiReceipt.Behaviours"
              xmlns:converters="clr-namespace:DigiReceipt.Converters"
              x:Class="DigiReceipt.ViewReceipts"
              x:Name="ViewReceiptsPage"
              Style="{StaticResource PageStyle}"
-             Title="{local:Translate txtViewReceiptsPageTitle}">
+             Title="{markup:Translate txtViewReceiptsPageTitle}">
     
     <ContentPage.ToolbarItems>
-        <ToolbarItem Text="{local:Translate txtLogoutButton}" Icon="Icons/logout.png" Clicked="OnLogout"/>
-        <ToolbarItem Text="{local:Translate txtAddReceiptButton}" Icon="Icons/add.png" Clicked="OnAddReceipt"/>
+        <ToolbarItem Text="{markup:Translate txtLogoutButton}" Icon="Icons/logout.png" Clicked="OnLogout"/>
+        <ToolbarItem Text="{markup:Translate txtAddReceiptButton}" Icon="Icons/add.png" Clicked="OnAddReceipt"/>
     </ContentPage.ToolbarItems>
 
     <ContentPage.Resources>
         <ResourceDictionary>
             <converters:ItemVisibilityEventArgsConverter x:Key="ItemVisibilityConverter"/>
-            <local:BooleanConverter x:Key="BooleanConverter"/>
+            <converters:BooleanConverter x:Key="BooleanConverter"/>
         </ResourceDictionary>
     </ContentPage.Resources>
 
@@ -30,7 +30,7 @@
             </Grid.RowDefinitions>
             
             <StackLayout Grid.Row="0" Margin="2" Orientation="Horizontal" HorizontalOptions="CenterAndExpand">
-                <Label Text="{local:Translate txtReceiptBeforeLabel}" VerticalOptions="CenterAndExpand"/>
+                <Label Text="{markup:Translate txtReceiptBeforeLabel}" VerticalOptions="CenterAndExpand"/>
                 
                 <DatePicker VerticalOptions="CenterAndExpand" Date="{Binding IssuedAfter, Mode=TwoWay}" MaximumDate="{x:Static sys:DateTime.Now}">
                     <DatePicker.Format>dd-MM-yyyy</DatePicker.Format>
@@ -38,9 +38,9 @@
             </StackLayout>
 
             <StackLayout Grid.Row="1">
-                <Label Text="{local:Translate txtLoadingLabel}" HorizontalOptions="CenterAndExpand" IsVisible="{Binding IsLoading}" FontSize="Medium" FontAttributes="Bold"/>
+                <Label Text="{markup:Translate txtLoadingLabel}" HorizontalOptions="CenterAndExpand" IsVisible="{Binding IsLoading}" FontSize="Medium" FontAttributes="Bold"/>
 
-                <Label Text="{local:Translate txtNoReceiptsLabel}" HorizontalOptions="CenterAndExpand" IsVisible="{Binding HasNoReceipts}"/>
+                <Label Text="{markup:Translate txtNoReceiptsLabel}" HorizontalOptions="CenterAndExpand" IsVisible="{Binding HasNoReceipts}"/>
 
                 <ListView Grid.Row="1" ItemsSource="{Binding Receipts, Mode=OneWay}" SelectedItem="{Binding SelectedItem, Mode=TwoWay}" ItemTapped="OnListViewItemTapped">
                     <ListView.Behaviors>
@@ -56,27 +56,27 @@
                                 <ViewCell.View>
                                     <StackLayout Orientation="Horizontal" HorizontalOptions="Fill">
                                         <!-- Display image of receipt -->
-                                        <Image Source="{Binding Image}" HeightRequest="50" WidthRequest="50" Aspect="AspectFit" BackgroundColor="Black" HorizontalOptions="Center" VerticalOptions="Center"/>
+                                        <Image Source="{Binding Image}" HeightRequest="50" WidthRequest="50" Aspect="AspectFit" BackgroundColor="Black"/>
 
                                         <!-- Display details of receipt -->
-                                        <StackLayout Orientation="Vertical">
+                                        <StackLayout Orientation="Vertical" HorizontalOptions="StartAndExpand">
                                             <StackLayout Orientation="Horizontal">
-                                                <Label Text="{local:Translate txtPriceLabel}"/>
+                                                <Label Text="{markup:Translate txtPriceLabel}"/>
                                                 <Label Text="{Binding Price, StringFormat='{0:C}'}"/>
                                             </StackLayout>
                                             <StackLayout Orientation="Horizontal">
-                                                <Label Text="{local:Translate txtIssuedOnLabel}"/>
+                                                <Label Text="{markup:Translate txtIssuedOnLabel}"/>
                                                 <Label Text="{Binding IssuedOn, StringFormat='{0:dd MMMM, yyyy}'}"/>
                                             </StackLayout>
                                         </StackLayout>
 
                                         <!-- Buttons-->
-                                        <StackLayout Orientation="Horizontal" HorizontalOptions="EndAndExpand" VerticalOptions="CenterAndExpand" Margin="0,0,10,0">
+                                        <StackLayout Orientation="Horizontal" VerticalOptions="CenterAndExpand" Margin="0,0,10,0">
                                             <!-- Edit -->
-                                            <Button Command="{Binding Source={x:Reference ViewReceiptsPage}, Path=BindingContext.EditReceiptCommand}" CommandParameter="{Binding}" Image="Icons/edit.png" BackgroundColor="Transparent" HorizontalOptions="End" VerticalOptions="Center"/>
+                                            <Button Command="{Binding Source={x:Reference ViewReceiptsPage}, Path=BindingContext.EditReceiptCommand}" CommandParameter="{Binding}" Image="Icons/edit.png" BackgroundColor="Transparent" HorizontalOptions="End" VerticalOptions="Center" WidthRequest="40" HeightRequest="40"/>
 
                                             <!-- Delete -->
-                                            <Button Command="{Binding Source={x:Reference ViewReceiptsPage}, Path=BindingContext.DeleteReceiptCommand}" CommandParameter="{Binding}" Image="Icons/delete.png" BackgroundColor="Transparent" HorizontalOptions="End" VerticalOptions="Center"/>
+                                            <Button Command="{Binding Source={x:Reference ViewReceiptsPage}, Path=BindingContext.DeleteReceiptCommand}" CommandParameter="{Binding}" Image="Icons/delete.png" BackgroundColor="Transparent" HorizontalOptions="End" VerticalOptions="Center" WidthRequest="40" HeightRequest="40"/>
                                         </StackLayout>
                                     </StackLayout>
                                 </ViewCell.View>
@@ -86,7 +86,7 @@
                 </ListView>
 
                 <!-- Offline message -->
-                <Label Text="{local:Translate txtOfflineLabel}" BackgroundColor="Black" TextColor="White" IsVisible="{Binding IsOnline, Converter={StaticResource BooleanConverter}}"/>
+                <Label Text="{markup:Translate txtOfflineLabel}" BackgroundColor="Black" TextColor="White" IsVisible="{Binding IsOnline, Converter={StaticResource BooleanConverter}}"/>
             </StackLayout>
         </Grid>
     </ContentPage.Content>
